requirements.txt
airflow-dbt==0.4.0
apache-airflow==2.10.2
apache-airflow-providers-google==10.24.0
dbt-bigquery==1.8.3
dbt-core==1.8.7
python-dotenv==1.0.1
